# docs/swagger.yaml
definitions:
  controllers.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  usecases.AuthUserInput:
    properties:
      email:
        example: juliano@test.com
        type: string
      password:
        example: P4ssw0rd!
        type: string
    required:
    - email
    - password
    type: object
  usecases.AuthUserOutput:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3MjA4Mjk5MDAsInN1YiI6IjUwNmU4Mjc4LTNiMzktNDI0ZS04NGU4LWMzYTE4NzcwNzBiNyJ9.AlUCK04QFIcGwlRw0e29fRMYlzZ3V979EH3pWlFVA1g
        type: string
    required:
    - token
    type: object
...
  usecases.RegisterUserInput:
    properties:
      email:
        example: juliano@test.com
        type: string
      name:
        example: Juliano
        type: string
      password:
        example: P4ssw0rd!
        type: string
    required:
    - email
    - name
    - password
    type: object
  usecases.RegisterUserOutput:
    properties:
      id:
        example: 506e8278-3b39-424e-84e8-c3a1877070b7
        type: string
    required:
    - id
    type: object
host: fasttrip.onrender.com
info:
  contact: {}
  description: Fasttrip API permite executar operações para cadastrar quartos, fazer
    reservas, checkin e checkout.
  title: Fasttrip API
  version: v1.7.1
paths:
  /auth_user:
    post:
      consumes:
      - application/json
      description: Autentica usuário
      parameters:
      - description: AuthUserInput
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecases.AuthUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.AuthUserOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Autentica usuário
      tags:
      - user
...
  /register_user:
    post:
      consumes:
      - application/json
      description: Registra novo usuário
      parameters:
      - description: RegisterUserInput
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecases.RegisterUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/usecases.RegisterUserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Registra novo usuário
      tags:
      - user
swagger: "2.0"